backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.1
django-cors-headers==4.4.0
djangorestframework-simplejwt==5.3.1
django-filter==24.2
drf-spectacular==0.27.1
dj-database-url==2.1.0
psycopg[binary]==3.1.18
python-dotenv==1.0.1
playwright==1.45.0
gunicorn==22.0.0

Pillow==10.4.0
